fix(agnocastlib): extract prepare_epoll as independent function
